fix identity manager Visual tests and add to the list (#16)
* fix identity manager Visual tests and add to the list

* add more screenshots

* wait for server ready when connecting to deployed app

* fix localization Visual tests and add to the list

* check that app is running before enabling features, and clean up

* fix quotes in NPM command when node is in .vaadin, improve logs

* add --cluster= parameter in order to run tests in any cluster

* set cluster context, and check https in installed apps

* fix kubectl port-forward in windows

* fix return when running playwrightTests in CC

* rename vars and check port 443

---------

+++ scripts/pit/lib/lib-ccenter.sh
@@ -14,12 +14,11 @@ CC_TLS_K=cc-control-login-tls
 ## Ingress names
 CC_ING_A=control-center
 CC_ING_K=control-center-keycloak-ingress
-## K8s cluster and namespace
-CC_CLUSTER=cc-cluster
+# Namespace used for CC
 CC_NS=control-center
 
 ## UI tests to run after the control-center is installed
-CC_TESTS="cc-setup.js cc-install-apps.js cc-identity-management.js"
+CC_TESTS=${CC_TESTS:-cc-setup.js cc-install-apps.js cc-identity-management.js cc-localization.js}
 
 checkDockerRunning() {
   if ! docker ps > /dev/null 2>&1; then
@@ -30,6 +29,7 @@ checkDockerRunning() {
 
 ## Install Control Center with Helm
 installCC() {
+  [ -n "SKIPHELM" ] && H=`kubectl get pods 2>&1` && echo "$H" | egrep -q 'control-center-[0-9abcdef]+-..... ' && return 0
   [ -n "$VERBOSE" ] && D=--debug || D=""
   [ -n "$CC_KEY" -a -n "$CC_CERT" ] && args="--set app.tlsSecret=$CC_TLS_A --set keycloak.tlsSecret=$CC_TLS_K" || args=""
   runCmd "$TEST" "Installing Vaadin Control Center" \
@@ -72,31 +72,46 @@ uninstallCC() {
   [ $? = 0 ] && echo "$H" | egrep -q "^$CC_NS " || return 0
   [ -n "$VERBOSE" ] && HD=--debug && KD=--v=10
   runCmd "$TEST" "Uninstalling Control-Center" helm uninstall control-center --wait -n $CC_NS $HD
-  runCmd "$TEST" "Removing namespace $CC_NS" kubectl delete ns $CC_NS $KD
+  runCmd "$TEST" "Removing namespace $CC_NS" kubectl delete ns $CC_NS $KD $1
+}
+
+getTLs() {
+  log "TLS config for ingress: $1"
+  H=`kubectl get ingress $1 -n $CC_NS -o jsonpath='{.spec.rules[0].host}'`
+  HS=`kubectl get ingress $1 -n $CC_NS -o jsonpath='{.spec.tls[*].hosts[*]}'`
+  S=`kubectl get ingress $1 -n $CC_NS -o jsonpath='{.spec.tls[*].secretName}'`
+  C=`kubectl get secret $S -n $CC_NS -o go-template='{{ index .data "tls.crt" | base64decode }}' | openssl x509 -noout -issuer -subject -enddate | tr '\n' ' '`
+  log "Host: $H is in ingress $1 with TLS config\n   hosts: $HS secret: $S cert: $C"
 }
 
 checkTls() {
   [ -n "$TEST" ] && return 0
+  log "Checking TLS certificates for all ingresses hosted in the cluster"
+  [ -n "$TEST" -o -n "$SKIPHELM" ] && return 0
   for i in `kubectl get ingresses -n $CC_NS | grep nginx | awk '{print $1}'`; do
-    log "$i"
-    H=`kubectl get ingress $i -n $CC_NS -o jsonpath='{.spec.rules[0].host}'`
-    HS=`kubectl get ingress $i -n $CC_NS -o jsonpath='{.spec.tls[*].hosts[*]}'`
-    S=`kubectl get ingress $i -n $CC_NS -o jsonpath='{.spec.tls[*].secretName}'`
-    C=`kubectl get secret $S -n $CC_NS -o go-template='{{ index .data "tls.crt" | base64decode }}' | openssl x509 -noout -issuer -subject -enddate | tr '\n' ' '`
-    log "Host: $H is in ingress $i with TLS config\n   hosts: $HS secret: $S cert: $C"
+    getTLs "$i"
   done
+}
+
+reloadIngress() {
+  [ -n "$TEST" ] && return 0
+  pod=`kubectl -n $CC_NS get pods | grep control-center-ingress-nginx-controller | awk '{print $1}'` || return 1
+  [ -n "$pod" ] && runCmd "$TEST" "Reloading nginx in $pod" "kubectl exec $pod -n "$CC_NS" -- nginx -s reload" || return 1
+  [ -z "$TEST" ] && sleep 3
 }
 
 ## Configure secrets for the control-center and the keycloak servers
 installTls() {
+  [ -n "$SKIPHELM" ] && return 0
   [ -z "$CC_KEY" -o -z "$CC_CERT" ] && log "No CC_KEY and CC_CERT provided, skiping TLS installation" && return 0
-  [ -n "$CC_FULL" ] && CC_CERT=$CC_FULL
-  [ -z "$TEST" ] && log "Installing TLS $CC_TLS for $CC_CONTROL and $CC_AUT" || cmd "## Creating TLS file '$CC_DOMAIN.pem' from envs"
+  # [ -n "$CC_FULL" ] && CC_CERT="$CC_FULL"
+  [ -z "$TEST" ] && log "Installing TLS $CC_TLS for $CC_CONTROL and $CC_AUTH" || cmd "## Creating TLS file '$CC_DOMAIN.pem' from envs"
   f1=cc-tls.crt
   f2=cc-tls.key
+  f3=$CC_DOMAIN.pem
   echo -e "$CC_CERT" > $f1 || return 1
   echo -e "$CC_KEY" > $f2 || return 1
-  cat $f1 $f2 > $CC_DOMAIN.pem
+  cat $f1 $f2 > $f3
 
   # remove old secrets if they exist (only needed for testing purposes since secrets are deleted before running the helm chart)
   kubectl get secret $CC_TLS_A -n $CC_NS >/dev/null 2>&1 && kubectl delete secret $CC_TLS_A -n $CC_NS
@@ -118,9 +133,7 @@ installTls() {
 
   [ -n "$TEST" ] && return 0
 
-  pod=`kubectl -n $CC_NS get pods | grep control-center-ingress-nginx-controller | awk '{print $1}'` || return 1
-  [ -n "$pod" ] && runCmd "$TEST" "Reloading nginx in $pod" "kubectl exec $pod -n "$CC_NS" -- nginx -s reload" || return 1
-  runCmd "$TEST" "Waiting for reloading ingress" sleep 5
+  reloadIngress || return 1
 }
 
 ## Show temporary user-email and password in the terminal
@@ -135,42 +148,73 @@ showTemporaryPassword() {
 ## Run Playwright tests for the control-center
 runPwTests() {
   computeNpm
+  [ -d screenshots.out ] && runCmd "$TEST" "Removing old screenshots" "rm -rf screenshots.out"
   [ -n "$SKIPPW" ] && return 0
   [ -z "$CC_CERT" -o -z "$CC_KEY" ] && NO_TLS=--notls || NO_TLS=""
   for f in $CC_TESTS; do
     runPlaywrightTests "$PIT_SCR_FOLDER/its/$f" "" "prod" "control-center" --url=https://$CC_CONTROL  --login=$CC_EMAIL $NO_TLS || return 1
-    sleep 3
+    if [ "$f" = cc-install-apps.js ]; then
+      reloadIngress && checkTls || return 1
+    fi
   done
+}
+
+setClusterContext() {
+  [ "$1" = "$KIND_CLUSTER" ] && current=kind-$1 || current=$1
+  ns=$2
+  H=`kubectl config get-contexts  | tr '*' ' ' | awk '{print $1}' | egrep "^$current$"`
+  [ -z "$H" ] && log "Cluster $current not found in kubectl contexts" && return 1
+  runCmd "$TEST" "Setting context to $current" "kubectl config use-context $current" || return 1
+  H=`kubectl config current-context`
+  [ "$H" != "$current" ] && log "Current context is not $current" && return 1
+  runCmd "$TEST" "Setting default namespace to $ns" "kubectl config set-context --current --namespace=$ns" || return 1
+  kubectl get ns >/dev/null 2>&1 || return 1
 }
 
 ## Main method for running control center
 runControlCenter() {
-  checkCommands kind helm docker kubectl || return 1
-  checkBusyPort "443" || return 1
+  CLUSTER=${CLUSTER:-$KIND_CLUSTER}
+
+  checkCommands docker kubectl helm unzip || return 1
   checkDockerRunning || return 1
-  ## Clean up from a previous run
-  # stopCloudProvider
-  uninstallCC $CC_CLUSTER $CC_NS
-  # deleteCluster $CC_CLUSTER
-  ## Start a new cluster
-  createCluster $CC_CLUSTER $CC_NS || return 1
-  # startCloudProvider || return 1
+
+  ## Start a new kind cluster if needed
+  [ "$CLUSTER" != "$KIND_CLUSTER" ] || createKindCluster $CLUSTER $CC_NS || return 1
+
+  ## Set the context to the cluster
+  setClusterContext "$CLUSTER" "$CC_NS" || return 1
+
+  ## Clean up CC from a previous run unless SKIPHELM is set
+  [ -z "$SKIPHELM" ] && uninstallCC
+
+  ## Check if port 443 is busy
+  checkBusyPort "443" || return 1
+
   ## Install Control Center
   installCC || return 1
   ## Control center takes a long time to start
   waitForCC 900 || return 1
+
   ## Show temporary user-email and password in the terminal
   showTemporaryPassword
+
   ## Install TLS certificates for the control-center and keycloak
   installTls && checkTls || return 1
+
   ## Forward the ingress (it needs root access since it uses port 443)
-  forwardIngress $CC_NS || return 1
-  ## Run Playwright tests
+  # checkPort "443"
+  [ "$CLUSTER" != "$KIND_CLUSTER" ] || forwardIngress $CC_NS || return 1
+
+  ## Run Playwright tests for the control-center
   runPwTests || return 1
-  if [ -z "$TEST" -a -z "$KEEPCC" ]; then
-    stopForwardIngress || return 1
-    deleteCluster $CC_CLUSTER || return 1
-  fi
+  stopForwardIngress || return 1
+
+  ## Delete the KinD cluster if it was created in this test if --keep-cc is not set
+  [ -n "$TEST" -o -n "$KEEPCC" -o "$CLUSTER" != "$KIND_CLUSTER" ] || deleteKindCluster "$CLUSTER" || return 1
+  ## Otherwise, uninstall the control-center if --keep-cc is not set
+  [ -n "$TEST" -o -n "$KEEPCC" -o "$CLUSTER"  = "$KIND_CLUSTER" ] || uninstallCC --wait=false || return 1
+
+  return 0
 }
